requirements.txt
numpy
easydict
pygame-ce>=2
vbao-mvvm